requirements1.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
deprecation==2.1.0
Flask==2.3.3
gotrue==1.3.1
gunicorn==21.2.0
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
postgrest==0.13.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
realtime==1.0.6
requests==2.31.0
six==1.17.0
sniffio==1.3.1
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.3
supafunc==0.3.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==12.0
Werkzeug==3.1.5
